working

+++ src/main/kotlin/com/lgaertner/euler/math/Max.kt
@@ -1,7 +1,7 @@
 package com.lgaertner.euler.math
 
-class Max(private val values: Wrapper<Collection<Int>>) {
-    fun value(): Int {
+class Max(private val values: Wrapper<Collection<Long>>) {
+    fun value(): Long {
         return values.collapse().max()
     }
 }
